test: tests using check_win_condition() from check_game_end.lua end faster
The big maps used for template testing have a maximum emulation speed of about 4.5x
on my machine. And the limit from the game is 65x.
The game seems to try to keep up to the missed time before ending. So a too high speed
delays ending and makes the whole run unnecessary long.

Try with a value of 20 (instead of 100) and see if this is better on other systems as well.

run time for test/templates/test_discovery-territorial_lord.lua on my machine:
before 48 sec (after 38 sec scripts are done)
now    41 sec